+++ src/main/java/com/tenio/core/api/implement/ServerApiImpl.java
@@ -104,7 +104,8 @@ public final class ServerApiImpl extends SystemLogger implements ServerApi {
     }
 
     try {
-      if (player.containsSession() && player.getSession().isPresent()) {
+      if (player.containsSession() && player.getSession().isPresent()
+              && player.getSession().get().isAssociatedToPlayer(Session.AssociatedState.DONE)) {
         // check process on method ZeroProcessorImpl#processSessionWillBeClosed
         Session session = player.getSession().get();
         session.close(connectionDisconnectMode, playerDisconnectMode);