+++ CHANGELOG.md
@@ -124,3 +124,4 @@
 ## 0.7.0 2026-04-12 Feature release
 - Supports Java 21 Features
 - Virtual Threads are being used in Session Inbound Process Queue and Schedulers
+- It should close sessions whenever exceptions occurred